+++ projects/ccf-eui/src/app/modules/filters/filters-content/filters-content.component.html
@@ -22,12 +22,6 @@
                 [selection]="filters?.tmc" (selectionChange)="updateFilter($event, 'tmc')"></ccf-checkbox>
 </div>
 
-<!-- <div class="button-container" [class.hidden]="hidden">
-  <button class="outline-button" mat-button (click)="applyButtonClick()">APPLY FILTERS</button>
-  <div class="refresh-icon">
-    <mat-icon class="icon refresh" (click)="refreshFilters()">refresh</mat-icon>
-  </div>
-</div> -->
 <div class="button-container">
   <button class="spatial-search-button" mat-button (click)="openSpatialSearch()">Spatial Search</button>
   <div class="button-container-end" [class.hidden]="hidden">
