+++ projects/ccf-eui/src/app/shared/components/spatial-search-config-behavior/spatial-search-config-behavior.component.html
@@ -1,1 +1,8 @@
-<ccf-spatial-search-config (updateSex)="updateSex($event)" (updateOrgan)="updateOrgan($event)" (closeDialog)="close()" (buttonClicked)="buttonClicked()" [organs]="filteredOrgans" [selectedOrgan]="selectedOrgan"></ccf-spatial-search-config>
+<ccf-spatial-search-config
+  (updateSex)="updateSex($event)"
+  (updateOrgan)="updateOrgan($event)"
+  (closeDialog)="close()"
+  (buttonClicked)="buttonClicked()"
+  [organs]="filteredOrgans"
+  [selectedOrgan]="selectedOrgan">
+</ccf-spatial-search-config>
