+++ projects/ccf-eui/src/app/shared/components/spatial-search-config-behavior/spatial-search-config-behavior.module.ts
@@ -1,7 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { SpatialSearchConfigModule } from '../spatial-search-config/spatial-search-config.module';
-import { MatDialogModule, MatDialogRef } from '@angular/material/dialog';
+import { MatDialogModule } from '@angular/material/dialog';
 
 import { SpatialSearchConfigBehaviorComponent } from './spatial-search-config-behavior.component';
 
@@ -10,12 +10,6 @@ import { SpatialSearchConfigBehaviorComponent } from './spatial-search-config-be
     CommonModule,
     SpatialSearchConfigModule,
     MatDialogModule
-  ],
-  providers: [
-    {
-      provide: MatDialogRef,
-      useValue: {}
-    }
   ],
   declarations: [SpatialSearchConfigBehaviorComponent],
   exports: [SpatialSearchConfigBehaviorComponent]
