+++ projects/ccf-eui/src/app/shared/components/spatial-search-config/spatial-search-config.module.ts
@@ -1,9 +1,9 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
+import { FormsModule } from '@angular/forms';
 import { MatSelectModule } from '@angular/material/select';
 import { MatIconModule } from '@angular/material/icon';
 import { MatRadioModule } from '@angular/material/radio';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
 
 import { SpatialSearchConfigComponent } from './spatial-search-config.component';
@@ -11,11 +11,10 @@ import { SpatialSearchConfigComponent } from './spatial-search-config.component'
 @NgModule({
   imports: [
     CommonModule,
+    FormsModule,
     MatSelectModule,
     MatIconModule,
     MatRadioModule,
-    FormsModule,
-    ReactiveFormsModule,
     MatButtonModule
   ],
   declarations: [SpatialSearchConfigComponent],
